training/requirements-train.txt
# Training dependencies (YOLO detector + TrOCR OCR fine-tuning)
ultralytics>=8.3.0
torch>=2.3.0
torchvision>=0.18.0
transformers>=4.44.0
accelerate>=0.33.0
datasets>=2.21.0
evaluate>=0.4.2
sentencepiece>=0.2.0
Pillow>=10.2.0
pandas>=2.1.0
numpy>=2.0.0
